exam can be assigned to user

+++ src/components/home/User.jsx
@@ -1,20 +1,26 @@
 import React from 'react';
+import Select from 'react-select';
 import { connect } from 'react-redux';
-import { setUser, setUsers, deleteUser } from '../../actions';
+import { setUser, setUsers, deleteUser, getExamSets, editUser } from '../../actions';
 
 class User extends React.Component {
 
   constructor(props) {
     super(props);
-    this.state = { email:"", name:"", phone:"", status:"", marks:"", timeStamp:""};
-    // console.log(props);
+    this.state = { email:"", name:"", phone:"", status:"", marks:"", timeStamp:"", user:"", examSet:""};
+
   }
 
-  componentWillMount() {
+  componentDidMount() {
+    this.props.getExamSets();
     this.props.setUsers();
   }
 
   renderUser() {
+
+    this.props.examSet.map(exam => exam.label = exam.name)
+    this.props.examSet.map(exam => exam.value = exam.set_id)
+
     if(this.props.users){
       let i = 0;
       return this.props.users.map(u=>{
@@ -25,20 +31,50 @@ class User extends React.Component {
             <td>{u.name}</td>
             <td>{u.email}</td>
             <td>{u.phone}</td>
-            <td>{u.status}</td>
-            <td>{u.marks}</td>
+            <td>
+              { (u.status==="Assigned")? u.status+", Exam :"+u.set_id : "" }
+            </td>
+            <td>
+              { u.marks? u.marks+", Exam : "+u.set_id :  "" }
+            </td>
             <td>{u.timeStamp}</td>
-            <td><button onClick={this.deleteUser.bind(this,i)}> Delete </button> </td>
+            <td>
+
+              <Select options={this.props.examSet}
+              onChange={opt => this.assignExamDropdown(opt, u)}/>
+
+              &nbsp;&nbsp;
+              <button onClick={this.assignExam}> Assign </button>
+              &nbsp;&nbsp;
+              <button onClick={this.deleteUser.bind(this, i)}> Delete User </button>
+
+            </td>
           </tr>
         )
       })
     }
   }
 
+  assignExamDropdown( opt, user ) {
+    console.log(opt.set_id);
+    console.log(user.user_id);
+    this.setState({
+      user: user,
+      examSet: opt
+    })
+  }
+
+  assignExam = ()=>{
+    // console.log(this.state);
+    if(window.confirm(" Are you sure to assign "+this.state.user.name+" the "+ this.state.examSet.name + " Exam set")){
+      this.props.editUser(this.state.user, this.state.examSet)
+    }
+  }
+
   deleteUser(i) {
-    // console.log(i);
-    // console.log(this.props);
-    this.props.deleteUser(i-1);
+    if(window.confirm("Are you sure to delete this user? ")){
+      this.props.deleteUser(i-1);
+    }
   }
 
   createNewUser = async()=>{
@@ -94,8 +130,8 @@ class User extends React.Component {
              <th scope="col">Phone</th>
              <th scope="col">Status</th>
              <th scope="col">Marks</th>
-             <th scope="col">Exam Time</th>
-             <th scope="col">Action</th>
+             <th scope="col">Exam Scheduled on</th>
+             <th scope="col">Actions</th>
            </tr>
          </thead>
          <tbody>
@@ -109,10 +145,10 @@ class User extends React.Component {
 
 const mapStateToProps = (state) => {
 
-  return { users: state.userReducer, questions: state.questionReducer};
+  return { users: state.userReducer, questions: state.questionReducer, examSet: state.examSetReducer};
 };
 
 export default connect(
   mapStateToProps,
-  { setUser, setUsers, deleteUser }
+  { setUser, setUsers, deleteUser, getExamSets, editUser }
 )(User);
